Add addons to wesnoth.game_config
This adds:
- wesnoth.game_config.eras
- wesnoth.game_config.campaigns
- wesnoth.game_config.modifications
- wesnoth.game_config.resources

Each is an ID to config mapping of the particular addon type.

wesnoth.game_config.eras previously existed as just an array of all era IDs, so this is an incompatible change. However, the old API was never documented, and the comment said it was specifically for @sigurdfd who approved this change.